requirements.txt
# PV Sentinel MVP - Python Dependencies
# Core Framework
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# AI/ML Libraries
torch>=2.1.0
transformers>=4.35.0
langchain>=0.0.350
langchain-community>=0.0.10
sentence-transformers>=2.2.2

# Vector Database & RAG
chromadb>=0.4.15
faiss-cpu>=1.7.4

# Audio Processing (Whisper)
openai-whisper>=20231117
librosa>=0.10.1
soundfile>=0.12.1
pyaudio>=0.2.11

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database & Storage
sqlite3
sqlalchemy>=2.0.23
alembic>=1.13.0

# Configuration & Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0

# PDF Generation & Export
reportlab>=4.0.7
jinja2>=3.1.2
weasyprint>=60.2

# Security & Authentication
bcrypt>=4.1.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Requests & APIs
requests>=2.31.0
httpx>=0.25.0

# Date/Time Handling
python-dateutil>=2.8.2
pytz>=2023.3

# Logging & Monitoring
loguru>=0.7.2
structlog>=23.2.0

# Testing & Quality
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1

# File Processing
python-magic>=0.4.27
chardet>=5.2.0

# Validation & Compliance
jsonschema>=4.20.0
cerberus>=1.3.5

# Development Tools
jupyterlab>=4.0.8
ipykernel>=6.26.0

# Production Deployment
gunicorn>=21.2.0
docker>=6.1.3

# Model Optimization
onnx>=1.15.0
onnxruntime>=1.16.3

# Medical Coding (MedDRA support)
# Custom packages would be added here for MedDRA integration

# Environment-specific packages
# Windows
pywin32>=306; sys_platform == "win32"

# macOS
pyobjc-framework-Cocoa>=10.0; sys_platform == "darwin"

# Linux
python3-dev>=3.8; sys_platform == "linux" 